% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakData.R
\name{as.peak}
\alias{as.peak}
\title{Coercion to PeakData Object
Converts numeric data to a PeakData object.}
\usage{
as.peak(positions, heights)
}
\arguments{
\item{positions}{Numeric vector of peak positions.}

\item{heights}{Numeric vector of peak heights.}
}
\value{
A PeakData object.
}
\description{
Coercion to PeakData Object
Converts numeric data to a PeakData object.
}
