% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakProminenceVisualizer.R
\name{as.peakVisualizer}
\alias{as.peakVisualizer}
\title{PeakProminenceVisualizer for 2D Data
Creates a PeakProminenceVisualizer object to handle peak and prominence visualization.}
\usage{
as.peakVisualizer(data)
}
\arguments{
\item{data}{A numeric matrix representing the 2D data.}
}
\value{
A list of PeakProminenceVisualizer objects for each row.
}
\description{
PeakProminenceVisualizer for 2D Data
Creates a PeakProminenceVisualizer object to handle peak and prominence visualization.
}
