% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakProminence.R
\name{as.prominence}
\alias{as.prominence}
\title{Coercion to PeakProminence Object
Converts numeric data to a PeakProminence object.}
\usage{
as.prominence(positions, prominences)
}
\arguments{
\item{positions}{Numeric vector of peak positions.}

\item{prominences}{Numeric vector of prominence values.}
}
\value{
A PeakProminence object.
}
\description{
Coercion to PeakProminence Object
Converts numeric data to a PeakProminence object.
}
