% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prominence.R
\name{calculate_prominence}
\alias{calculate_prominence}
\title{Calculate Prominence for Peaks}
\usage{
calculate_prominence(data, peaks)
}
\arguments{
\item{data}{A numeric vector representing the 2D data.}

\item{peaks}{A Peak object containing detected peaks.}
}
\value{
A Prominence object containing the prominences of each peak.
}
\description{
Calculate Prominence for Peaks
}
