% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_info.R
\name{get_peak_info}
\alias{get_peak_info}
\title{Get combined peak information as a data frame}
\usage{
get_peak_info(peaks, prominence)
}
\arguments{
\item{peaks}{A Peak object.}

\item{prominence}{A Prominence object.}
}
\value{
A data frame with Peak Position, Peak Height, and Prominence.
}
\description{
Get combined peak information as a data frame
}
