% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_class.R
\name{Peak}
\alias{Peak}
\title{Peak Class Constructor
Creates a Peak object containing peak positions and heights.}
\usage{
Peak(positions, heights)
}
\arguments{
\item{positions}{Numeric vector of peak positions.}

\item{heights}{Numeric vector of peak heights.}
}
\value{
A Peak object.
}
\description{
Peak Class Constructor
Creates a Peak object containing peak positions and heights.
}
