% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_visualizer_class.R
\name{PeakVisualizer}
\alias{PeakVisualizer}
\title{PeakVisualizer Class Constructor
Creates a PeakVisualizer object to handle peak and prominence visualization.}
\usage{
PeakVisualizer(data)
}
\arguments{
\item{data}{Numeric vector of data points.}
}
\value{
A PeakVisualizer object.
}
\description{
PeakVisualizer Class Constructor
Creates a PeakVisualizer object to handle peak and prominence visualization.
}
