% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visualization.R
\name{plot_peaks.PeakVisualizer}
\alias{plot_peaks.PeakVisualizer}
\title{Plot Peaks}
\usage{
plot_peaks.PeakVisualizer(object)
}
\arguments{
\item{object}{A PeakVisualizer object.}
}
\description{
Plot Peaks
}
