% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_visualizer_class.R
\name{plot_peaks}
\alias{plot_peaks}
\title{Plot Peaks for PeakVisualizer}
\usage{
plot_peaks(object)
}
\arguments{
\item{object}{A PeakVisualizer object.}
}
\description{
Plot Peaks for PeakVisualizer
}
