% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visualization.R
\name{plot_prominence.PeakVisualizer}
\alias{plot_prominence.PeakVisualizer}
\title{Plot Prominence}
\usage{
plot_prominence.PeakVisualizer(object)
}
\arguments{
\item{object}{A PeakVisualizer object.}
}
\description{
Plot Prominence
}
