% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_visualizer_class.R
\name{plot_prominence}
\alias{plot_prominence}
\title{Plot Prominence for PeakVisualizer}
\usage{
plot_prominence(object)
}
\arguments{
\item{object}{A PeakVisualizer object.}
}
\description{
Plot Prominence for PeakVisualizer
}
