% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakData.R
\name{print.PeakData}
\alias{print.PeakData}
\title{Print Method for PeakData Class}
\usage{
\method{print}{PeakData}(x, ...)
}
\arguments{
\item{x}{A PeakData object.}

\item{...}{Additional parameters (currently unused).}
}
\description{
Print Method for PeakData Class
}
