% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakProminence.R
\name{print.PeakProminence}
\alias{print.PeakProminence}
\title{Print Method for PeakProminence Class}
\usage{
\method{print}{PeakProminence}(x, ...)
}
\arguments{
\item{x}{A PeakProminence object.}

\item{...}{Additional parameters.}
}
\description{
Print Method for PeakProminence Class
}
