% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prominence_class.R
\name{print.Prominence}
\alias{print.Prominence}
\title{Print method for Prominence objects}
\usage{
\method{print}{Prominence}(x, ...)
}
\arguments{
\item{x}{A Prominence object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for Prominence objects
}
