% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prominence_class.R
\name{Prominence}
\alias{Prominence}
\title{Prominence Class Constructor
Creates a Prominence object.}
\usage{
Prominence(positions, prominences)
}
\arguments{
\item{positions}{Numeric vector of peak positions.}

\item{prominences}{Numeric vector of prominence values.}
}
\value{
A Prominence object.
}
\description{
Prominence Class Constructor
Creates a Prominence object.
}
