% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_class.R
\name{summary.Peak}
\alias{summary.Peak}
\title{Summary Method for Peak Class}
\usage{
\method{summary}{Peak}(object, ...)
}
\arguments{
\item{object}{A Peak object.}

\item{...}{Additional parameters.}
}
\description{
Summary Method for Peak Class
}
