% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakData.R
\name{summary.PeakData}
\alias{summary.PeakData}
\title{Summary Method for PeakData Class}
\usage{
\method{summary}{PeakData}(object, ...)
}
\arguments{
\item{object}{A PeakData object.}

\item{...}{Additional parameters.}
}
\description{
Summary Method for PeakData Class
}
