% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakProminence.R
\name{summary.PeakProminence}
\alias{summary.PeakProminence}
\title{Summary Method for PeakProminence Class}
\usage{
\method{summary}{PeakProminence}(object, ...)
}
\arguments{
\item{object}{A PeakProminence object.}

\item{...}{Additional parameters.}
}
\description{
Summary Method for PeakProminence Class
}
