% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prominence.R
\name{calculate_prominence}
\alias{calculate_prominence}
\title{Calculate Prominence for Detected Peaks}
\usage{
calculate_prominence(peaks, data)
}
\arguments{
\item{peaks}{A Peak object containing peak positions and heights.}

\item{data}{A numeric vector representing the original data.}
}
\value{
A Prominence object with calculated prominence values for each peak.
}
\description{
Calculate Prominence for Detected Peaks
}
