% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_peaks.R
\name{detect_peaks}
\alias{detect_peaks}
\title{Detect Peaks in 2D Data
This function uses \code{pracma::findpeaks} to detect peaks in a numeric vector.}
\usage{
detect_peaks(data)
}
\arguments{
\item{data}{A numeric vector representing the data.}
}
\value{
A Peak object with positions and heights of detected peaks.
}
\description{
Detect Peaks in 2D Data
This function uses \code{pracma::findpeaks} to detect peaks in a numeric vector.
}
