% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_peaks.R
\name{detect_peaks}
\alias{detect_peaks}
\title{Detect Peaks in 2D Data}
\usage{
detect_peaks(data, min_peak_height = 5)
}
\arguments{
\item{data}{A numeric vector representing the 2D data.}

\item{min_peak_height}{The minimum height a peak must have to be considered a peak.}
}
\value{
A Peak object containing the positions and heights of the detected peaks.
}
\description{
Detect Peaks in 2D Data
}
