% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_class.R
\name{Peak}
\alias{Peak}
\title{Create a Peak object}
\usage{
Peak(positions, heights)
}
\arguments{
\item{positions}{A numeric vector of peak positions.}

\item{heights}{A numeric vector of peak heights.}
}
\value{
An object of class Peak containing the positions and heights of detected peaks.
}
\description{
Create a Peak object
}
