% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_visualizer_class.R
\name{PeakVisualizer}
\alias{PeakVisualizer}
\title{PeakVisualizer class}
\usage{
PeakVisualizer(data, peaks, prominence = NULL)
}
\arguments{
\item{data}{A numeric vector representing the signal data.}

\item{peaks}{A Peak object containing detected peaks.}

\item{prominence}{(Optional) A Prominence object containing prominence values for the peaks.}
}
\value{
A PeakVisualizer object.
}
\description{
PeakVisualizer class
}
