% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visualization.R
\name{plot_peaks.PeakVisualizer}
\alias{plot_peaks.PeakVisualizer}
\title{Plot Peaks on 2D Data}
\usage{
plot_peaks.PeakVisualizer(visualizer)
}
\arguments{
\item{visualizer}{A PeakVisualizer object.}
}
\description{
Plot Peaks on 2D Data
}
