% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_visualizer_class.R
\name{plot_peaks.PeakVisualizer}
\alias{plot_peaks.PeakVisualizer}
\title{Plot detected peaks using PeakVisualizer}
\usage{
plot_peaks.PeakVisualizer(obj)
}
\arguments{
\item{obj}{A PeakVisualizer object.}
}
\description{
Plot detected peaks using PeakVisualizer
}
