% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visualization.R
\name{plot_peaks}
\alias{plot_peaks}
\title{Plot Detected Peaks for 1D or 2D Data
Visualizes the detected peaks for the entire data.}
\usage{
plot_peaks(
  peaks,
  data,
  main_title = "Peak Detection",
  peak_color = "red",
  line_color = "black"
)
}
\arguments{
\item{peaks}{A PeakData object (for 1D data) or a list of PeakData objects (for 2D data).}

\item{data}{A numeric vector (1D) or matrix (2D) representing the data.}

\item{main_title}{Title of the plot. Defaults to "Peak Detection".}

\item{peak_color}{Color for the peaks. Defaults to "red".}

\item{line_color}{Color for the line plot. Defaults to "black".}
}
\description{
Plot Detected Peaks for 1D or 2D Data
Visualizes the detected peaks for the entire data.
}
