% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_visualizer_class.R
\name{plot_prominence.PeakVisualizer}
\alias{plot_prominence.PeakVisualizer}
\title{Plot peak prominence using PeakVisualizer}
\usage{
plot_prominence.PeakVisualizer(obj)
}
\arguments{
\item{obj}{A PeakVisualizer object.}
}
\description{
Plot peak prominence using PeakVisualizer
}
