% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visualization.R
\name{plot_prominence.PeakVisualizer}
\alias{plot_prominence.PeakVisualizer}
\title{Plot Peaks and Prominences on 2D Data}
\usage{
plot_prominence.PeakVisualizer(visualizer)
}
\arguments{
\item{visualizer}{A PeakVisualizer object.}
}
\description{
Plot Peaks and Prominences on 2D Data
}
