% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visualization.R
\name{plot_prominence}
\alias{plot_prominence}
\title{Plot Peak Prominence for 1D or 2D Data
Visualizes the prominence of detected peaks for the entire data.}
\usage{
plot_prominence(
  peaks,
  prominence,
  data,
  main_title = "Prominence Detection",
  peak_color = "red",
  line_color = "black",
  prominence_color = "blue"
)
}
\arguments{
\item{peaks}{A PeakData object (for 1D data) or a list of PeakData objects (for 2D data).}

\item{prominence}{A Prominence object (for 1D data) or a list of Prominence objects (for 2D data).}

\item{data}{A numeric vector (1D) or matrix (2D) representing the data.}

\item{main_title}{Title of the plot. Defaults to "Prominence Detection".}

\item{peak_color}{Color for the peaks. Defaults to "red".}

\item{line_color}{Color for the line plot. Defaults to "black".}

\item{prominence_color}{Color for the prominence bars. Defaults to "blue".}
}
\description{
Plot Peak Prominence for 1D or 2D Data
Visualizes the prominence of detected peaks for the entire data.
}
