% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_class.R
\name{print.Peak}
\alias{print.Peak}
\title{Print Method for Peak Class}
\usage{
\method{print}{Peak}(x, ...)
}
\arguments{
\item{x}{A Peak object.}

\item{...}{Additional parameters.}
}
\description{
Print Method for Peak Class
}
