% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prominence_class.R
\name{Prominence}
\alias{Prominence}
\title{Create a Prominence object}
\usage{
Prominence(positions, prominences)
}
\arguments{
\item{positions}{A numeric vector of peak positions.}

\item{prominences}{A numeric vector of the prominence values for each peak.}
}
\value{
An object of class Prominence containing the positions and prominence values of the peaks.
}
\description{
Create a Prominence object
}
