% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_class.R
\name{summary.Peak}
\alias{summary.Peak}
\title{Summary method for Peak objects}
\usage{
\method{summary}{Peak}(object, ...)
}
\arguments{
\item{object}{A Peak object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for Peak objects
}
