% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prominence_class.R
\name{summary.Prominence}
\alias{summary.Prominence}
\title{Summary method for Prominence objects}
\usage{
\method{summary}{Prominence}(object, ...)
}
\arguments{
\item{object}{A Prominence object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for Prominence objects
}
